unstable commit: trying to make multi PK support work as table constraint.

+++ src/demo3/Payment.java
@@ -2,12 +2,14 @@ package demo3;
 
 import org.orman.mapper.Model;
 import org.orman.mapper.annotation.Entity;
+import org.orman.mapper.annotation.Index;
 import org.orman.mapper.annotation.PrimaryKey;
+import org.orman.sql.IndexType;
 
 @Entity
 public class Payment extends Model<Payment>{
 	@PrimaryKey(autoIncrement=true)
-	public long tx_id;
+	public int tx_id;
 	
 	public float amount;
